Made maxTime double in GameData::calculateRoundPoints to fix the calculation error

+++ Server/src/common/infrastructure/GameData.cpp
@@ -84,7 +84,7 @@ void GameData::calculateRoundPoints(const bool didFail)
         return;
     }
 
-    const int maxTime = this->game.getRoom().getData().timePerQuestionSecs * 1000;
+    const double maxTime = static_cast<double>(this->game.getRoom().getData().timePerQuestionSecs * 1000);
     const double time = static_cast<double>(this->roundTime.count());
 
     const double result =
